% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table1_utils_r_table.R
\name{r_table}
\alias{r_table}
\title{reactTable Template}
\usage{
r_table(table, key, bg)
}
\arguments{
\item{table}{Unfiltered datatable from}

\item{key}{Filter key, will be index in practice}

\item{bg}{background color specs}
}
\description{
reactTable Template
}
